Correction warnings Ubuntu

+++ src/Core/Map/Layers/CollisionLayer.cpp
@@ -27,7 +27,7 @@ void CollisionLayer::AddCollisionBox(const CollisionBox& newColBox){
 }
 
 void CollisionLayer::DisplayTXT() const{
-    for(int i = 0; i<collisionBoxes.size(); i++){
+    for(long unsigned int i = 0; i<collisionBoxes.size(); i++){
         std::cout << "Id : " << collisionBoxes[i].GetId() << std::endl;
         std::cout << "X : " << collisionBoxes[i].GetX() << std::endl;
         std::cout << "Y : " << collisionBoxes[i].GetY() << std::endl;
